Update SponsorUtility.R
Changes: 

The main change I did was to split the integration of S1RF and S2RF, which was in the previouse version a single integral, into 2 separate integrals. It appears that this is easier to handle for the R integrate() function. 

For R, integrals where the integrand is positive are easier to handle (e.g. IntegrandS1RF, which is the product of the density and 1- cdf. is always positive since the factors are always positive). 
This is different for the integrands
IntegrandS2RF and IntegrandS3RF. They are of the form zvalue * densitiy * cdf, and depending on zvalue, which might be negative, the integrand is for some zvalues negative. Integrals which are not positive over the entire domain are harder to integrate, sometimes R thinks the integrals are divergent and an "Error: integral might be divergent" is thrown. I did the following workarround based on the 
TryChatch function in R:
Assume an error is thrown when we want to integrate IntegrandS3RF. Then I check the integrals abs(IntegrandS3RF) and -abs(IntegrandS3RF). If both integrals over abs(IntegrandS3RF) and -abs(IntegrandS3RF) are very close to 0, we can be sure that IntegrandS3RF is close to zero. (This is due to the fact that -abs(IntegrandS3RF)<= IntegrandS3RF <= abs(IntegrandS3RF) and hence for the same inequalties hold true if you integrate over these three functions. If both integral(-abs(IntegrandS3RF)) and integral(abs(IntegrandS3RF)) are close to zero, then integral(IntegrandS3RF) will also be close to zero).

In addition I have changed the rel.tol of the integration rountines to smaller values (note that, for some numeric reason, too small rel.tol values also lead to an error, namly "Error: Error in extrapoloation table", I assume that this happens if the accuracy required for the integral exeeds the accuracy provided by the funtions pnorm, qnorm,..etc, but I am not completely sure about that). 

The rel.tols I have chosen are between .Machine$double.eps^0.75 and .Machine$double.eps ^0.5, these are just values which worked in practice, no rational is behind that.

The changes I made reduce the errors, but there is no guarantee that it will always work (e.g. if .Machine$double.eps^0.75 is asked too much an extrapolation error will be thrown, I am still working on integrating this in the code, however the current version might alredy suffice for our needs.)