Add files via upload
Added option to perform a "Grid" or "Optim" optimisation in the second stage. The "Grid" option performs optimisation over a discrete grid of sample sizes. This avoids the possible problem of local optima when using optim.

The stage 1 optimisation has also been updated. The summation approach for stage 1 and stage 2 are now very similar. A stage 2 optimal policy will only be constructed for the optimal stage 1 design.